fix: move oidc to Veramo.service, fixes to e2e tests
