+++ packages/snap/tests/e2e/createVerifiablePresentation.spec.ts
@@ -5,7 +5,7 @@ import type { SnapsGlobalObject } from '@metamask/snaps-types';
 import { VerifiablePresentation } from '@veramo/core';
 
 import { onRpcRequest } from '../../src';
-import { getAgent, type Agent } from '../../src/veramo/setup';
+import VeramoService, { type Agent } from '../../src/veramo/Veramo.service';
 import { account } from '../data/constants';
 import { getDefaultSnapState } from '../data/defaultSnapState';
 import exampleVCEIP712 from '../data/verifiable-credentials/exampleEIP712.json';
@@ -54,10 +54,11 @@ describe('createVerifiablePresentation', () => {
         newState: getDefaultSnapState(account),
       });
       snapMock.rpcMocks.snap_dialog.mockReturnValue(true);
-      const ethereumMock = snapMock as unknown as MetaMaskInpageProvider;
-      agent = await getAgent(snapMock, ethereumMock);
       global.snap = snapMock;
       global.ethereum = snapMock as unknown as MetaMaskInpageProvider;
+
+      await VeramoService.init();
+      agent = VeramoService.getAgent();
 
       const switchMethod = (await onRpcRequest({
         origin: 'localhost',
