+++ packages/snap/tests/e2e/saveVerifiableCredential.spec.ts
@@ -11,7 +11,7 @@ import type { IIdentifier, VerifiableCredential } from '@veramo/core';
 
 import { onRpcRequest } from '../../src';
 import type { StoredCredentials } from '../../src/veramo/plugins/ceramicDataStore/ceramicDataStore';
-import { getAgent, type Agent } from '../../src/veramo/setup';
+import VeramoService, { type Agent } from '../../src/veramo/Veramo.service';
 import { account, importablePrivateKey } from '../data/constants';
 import examplePayload from '../data/credentials/examplePayload.json';
 import { getDefaultSnapState } from '../data/defaultSnapState';
@@ -64,8 +64,6 @@ describe('saveVerifiableCredential', () => {
       operation: 'update',
       newState: getDefaultSnapState(account),
     });
-    global.snap = snapMock;
-    global.ethereum = snapMock as unknown as MetaMaskInpageProvider;
 
     // Clear stores before each test
     await agent.clear({ options: { store: ['snap', 'ceramic'] } });
@@ -78,12 +76,15 @@ describe('saveVerifiableCredential', () => {
       newState: getDefaultSnapState(account),
     });
     snapMock.rpcMocks.snap_dialog.mockReturnValue(true);
-    const ethereumMock = snapMock as unknown as MetaMaskInpageProvider;
-    agent = await getAgent(snapMock, ethereumMock);
+    global.snap = snapMock;
+    global.ethereum = snapMock as unknown as MetaMaskInpageProvider;
+
+    agent = await VeramoService.createAgent();
     identifier = await agent.didManagerCreate({
       provider: 'did:ethr',
       kms: 'snap',
     });
+
     await agent.keyManagerImport(importablePrivateKey);
 
     // Create test VC
